feat(commands): DetermineVersion and Standardized Version.json scripts (#4)
* Added some more useful scripts

determine-version can be added to post-commit hooks to display what version to update parent repositories too. generate-version-file should be added as a post-build step to every repository so that it generates the standardized version.json file. Still need to add dependencies to the version.json file

* Updates to style of messages logged at end of determine-version

* Include dependencies of repo too

Maybe include package.json of repo in its version.json?

* Use Jenkins multibranch pipeline environment var for branch

Need to do this because multibranch pipeline jobs check out commits not branches.

* Add -l param

* add location param

* debugging

* more debug logs

* fix location param

* Dont error if it cant determine the branch

* Update the README

* Nullhosn mistake

* README updates